// src/tournament-operation/types/tournament/TournamentFormat.ts
export enum TournamentFormat {
  LEAGUE = 'LEAGUE',
  GROUP_KNOCKOUT = 'GROUP_KNOCKOUT',
  KNOCKOUT = 'KNOCKOUT'
}

export enum TournamentPhase {
  GROUP_STAGE = 'GROUP_STAGE',
  KNOCKOUT_STAGE = 'KNOCKOUT_STAGE'
}

export enum MatchFormat {
  SINGLE_MATCH = 'SINGLE_MATCH',
  HOME_AWAY = 'HOME_AWAY'
}

export enum KnockoutQualification {
  POINTS = 'POINTS',
  GROUP_POSITION = 'GROUP_POSITION'
} 